refactored a bit

+++ Test Tasks/servermechanics.cz/Fibonacci.java
@@ -15,13 +15,13 @@ class Fibonacci {
     void computeAndShow(int count) {
         if (count < 2)
             return;
-        int total = 2;
+        float total = 2;
         System.out.print("1 1 ");
         for (int i = 3, n0 = 1, n1 = 1, n2 = 2; i <= count;
                 i++, n0 = n1, n1 = n2, n2 = n0 + n1) {
             total += n2;
             System.out.print(n2 + " ");
         }
-        System.out.println("\nThe average is " + (float)total / count);
+        System.out.println("\nThe average is " + total / count);
     }
 }
